A cleanup and improvements here and there
