+++ .env.example
@@ -3,3 +3,5 @@ KOL_PASS='Cagebot P4ssw0rD'
 MAINTAIN_ADVENTURES='80'
 OPEN_EVERYTHING='true/false'
 ONLY_OPEN_WHEN_ADVS_ABOVE='80'
+WHITEBOARD_CAGED='${name} (#${id}) is CAGED in Hobopolis'
+WHITEBOARD_UNCAGED='Hobopolis is UNCAGED'
